Extract DebounceComputer as a separate class
