+++ backend/core/src/main/java/org/sonarsource/sonarlint/core/commons/SmartCancelableLoadingCache.java
@@ -21,11 +21,9 @@ package org.sonarsource.sonarlint.core.commons;
 
 import com.google.common.util.concurrent.MoreExecutors;
 import java.util.concurrent.CancellationException;
-import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
 import java.util.function.BiFunction;
 import javax.annotation.Nullable;
 import org.sonarsource.sonarlint.core.commons.log.SonarLintLogger;
@@ -43,103 +41,7 @@ public class SmartCancelableLoadingCache<K, V> implements AutoCloseable {
   private final ExecutorServiceShutdownWatchable<?> executorService;
   private final String threadName;
   private final BiFunction<K, SonarLintCancelMonitor, V> valueComputer;
-  private final ConcurrentHashMap<K, ValueAndComputeFutures> cache = new ConcurrentHashMap<>();
-
-  private class ValueAndComputeFutures {
-    private final K key;
-    private CompletableFuture<V> valueFuture = new CompletableFuture<>();
-    private CompletableFuture<V> computeFuture;
-    private final ReentrantReadWriteLock lock = new ReentrantReadWriteLock();
-
-    public ValueAndComputeFutures(K key) {
-      this.key = key;
-    }
-
-    private void scheduleComputation() {
-      lock.writeLock().lock();
-      try {
-        if (computeFuture != null) {
-          computeFuture.cancel(false);
-          computeFuture = null;
-        }
-        // Even if we have canceled the previous computation, the previous valueFuture can still be completed later, so we will create a new one
-        // that will be returned to new callers. We will also try to complete the old valueFuture, in case the cancellation was successful.
-        var oldValueFuture = valueFuture;
-        valueFuture = new CompletableFuture<>();
-
-        var previousValue = oldValueFuture.getNow(null);
-
-        var cancelMonitor = new SonarLintCancelMonitor();
-        cancelMonitor.watchForShutdown(executorService);
-        CompletableFuture<V> newComputeFuture = CompletableFuture.supplyAsync(() -> {
-          cancelMonitor.checkCanceled();
-          return valueComputer.apply(key, cancelMonitor);
-        }, executorService);
-        newComputeFuture.whenComplete((newValue, error) -> {
-          if (error instanceof CancellationException) {
-            cancelMonitor.cancel();
-          }
-        });
-        newComputeFuture.whenCompleteAsync((newValue, error) -> whenComputeCompleted(newValue, error, previousValue, oldValueFuture), executorService);
-        computeFuture = newComputeFuture;
-      } finally {
-        lock.writeLock().unlock();
-      }
-    }
-
-    private void whenComputeCompleted(@Nullable V newValue, @Nullable Throwable error, @Nullable V previousValue, CompletableFuture<V> oldValueFuture) {
-      lock.writeLock().lock();
-      try {
-        computeFuture = null;
-        if (error instanceof CancellationException) {
-          return;
-        }
-        try {
-          if (error == null && listener != null) {
-            listener.afterCachedValueRefreshed(key, previousValue, newValue);
-          }
-        } finally {
-          if (error != null) {
-            oldValueFuture.completeExceptionally(error);
-            valueFuture.completeExceptionally(error);
-          } else {
-            oldValueFuture.complete(newValue);
-            valueFuture.complete(newValue);
-          }
-        }
-      } finally {
-        lock.writeLock().unlock();
-      }
-    }
-
-
-    public void refresh() {
-      scheduleComputation();
-    }
-
-    public CompletableFuture<V> getValueFuture() {
-      lock.readLock().lock();
-      try {
-        return valueFuture;
-      } finally {
-        lock.readLock().unlock();
-      }
-    }
-
-    public void cancel() {
-      lock.writeLock().lock();
-      try {
-        if (computeFuture != null) {
-          computeFuture.cancel(false);
-          computeFuture = null;
-        }
-        valueFuture.cancel(false);
-      } finally {
-        lock.writeLock().unlock();
-      }
-    }
-
-  }
+  private final ConcurrentHashMap<K, DebounceComputer<V>> cache = new ConcurrentHashMap<>();
 
   @Nullable
   private final Listener<K, V> listener;
@@ -179,7 +81,7 @@ public class SmartCancelableLoadingCache<K, V> implements AutoCloseable {
    */
   public void refreshAsync(K key) {
     cache.compute(key, (k, v) -> {
-      if (v == null)  {
+      if (v == null) {
         return newValueAndScheduleComputation(k);
       } else {
         v.refresh();
@@ -193,8 +95,12 @@ public class SmartCancelableLoadingCache<K, V> implements AutoCloseable {
     return resultFuture.getValueFuture().join();
   }
 
-  private ValueAndComputeFutures newValueAndScheduleComputation(K k) {
-    var value = new ValueAndComputeFutures(k);
+  private DebounceComputer<V> newValueAndScheduleComputation(K k) {
+    var value = new DebounceComputer<>(c -> valueComputer.apply(k, c), executorService, (oldValue, newValue) -> {
+      if (listener != null) {
+        listener.afterCachedValueRefreshed(k, oldValue, newValue);
+      }
+    });
     value.scheduleComputation();
     return value;
   }
